requirements.txt
fastapi
uvicorn
aiohttp
pillow
python-dotenv
supabase
httpx[http2]
jose
